+++ src/drawer/hooks/useGetDrawerBanner.ts
@@ -1,0 +1,12 @@
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+import { getDrawerBanner } from '../apis/getDrawerBanner';
+
+export const useGetDrawerBanner = () => {
+  return useSuspenseQuery({
+    queryKey: ['drawerBanner'],
+    queryFn: getDrawerBanner,
+    staleTime: Infinity,
+    select: (data) => data.bannerList,
+  });
+};
