+++ src/drawer/hooks/useGetStarRank.ts
@@ -1,4 +1,4 @@
-import { InfiniteData, useInfiniteQuery } from '@tanstack/react-query';
+import { InfiniteData, useSuspenseInfiniteQuery } from '@tanstack/react-query';
 
 import { getRanking } from '../apis/getRanking';
 import { PRODUCTS_PER_PAGE } from '../constants/page.constant';
@@ -8,7 +8,7 @@ import { ProductResponses, ProductResult } from '../types/product.type';
 type ProductRequestParamsWithoutPage = Omit<ProductRequestParams, 'page'>;
 
 export const useGetStarRank = ({ responseType, category }: ProductRequestParamsWithoutPage) => {
-  return useInfiniteQuery<
+  return useSuspenseInfiniteQuery<
     ProductResponses[],
     Error,
     InfiniteData<ProductResult[], number>,
