+++ src/drawer/types/banner.type.ts
@@ -1,0 +1,11 @@
+interface Banner {
+  id: number;
+  image: string;
+  name: string;
+  type: 'WEB' | 'APP' | 'WEBAPP';
+  website: string;
+}
+
+export interface DrawerBannerResponse {
+  bannerList: Banner[];
+}
